refactor: Move wagmi config to config/index.ts
